Move theme routes into ept-themes
