+++ modules/ept-themes/routes/index.js
@@ -1,0 +1,8 @@
+var path = require('path');
+
+module.exports = [
+  require(path.normalize(__dirname + '/getTheme')),
+  require(path.normalize(__dirname + '/previewTheme')),
+  require(path.normalize(__dirname + '/resetTheme')),
+  require(path.normalize(__dirname + '/setTheme'))
+];
