--- server/routes/admin/settings/index.js
@@ -1,9 +1,0 @@
-var path = require('path');
-var settings = require(path.normalize(__dirname + '/config'));
-
-module.exports = [
-  { method: 'GET', path: '/settings/theme', config: settings.getTheme },
-  { method: 'PUT', path: '/settings/theme', config: settings.setTheme },
-  { method: 'PUT', path: '/settings/theme/preview', config: settings.previewTheme },
-  { method: 'POST', path: '/settings/theme', config: settings.resetTheme }
-];
